+++ tracker/fingerprint.go
@@ -5,6 +5,7 @@ import (
 	"net"
 	"net/http"
 	"strings"
+	"time"
 
 	"github.com/dchest/siphash"
 )
@@ -14,12 +15,13 @@ var (
 )
 
 // Fingerprint returns a hash for given request and salt.
-// The hash is unique for the visitor.
-func Fingerprint(r *http.Request, salt string, headerParser []ip.HeaderParser, allowed []net.IPNet) uint64 {
+// The hash is unique for the visitor and day.
+func Fingerprint(r *http.Request, salt string, date time.Time, headerParser []ip.HeaderParser, allowed []net.IPNet) uint64 {
 	var sb strings.Builder
 	sb.WriteString(r.Header.Get("User-Agent"))
 	sb.WriteString(ip.Get(r, headerParser, allowed))
 	sb.WriteString(salt)
+	sb.WriteString(date.Format("20060102"))
 	return siphash.Hash(hashKey0, hashKey1, []byte(sb.String()))
 }