+++ tracker/fingerprint_test.go
@@ -4,6 +4,7 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 )
@@ -13,5 +14,5 @@ func TestFingerprint(t *testing.T) {
 	req := httptest.NewRequest(http.MethodGet, "/", nil)
 	req.Header.Set("User-Agent", "test")
 	req.RemoteAddr = "127.0.0.1:80"
-	assert.Equal(t, uint64(0x4f97de4b2cbf6e12), Fingerprint(req, "salt", nil, nil))
+	assert.Equal(t, uint64(0x5156f792eb1ac0fb), Fingerprint(req, "salt", time.Now().UTC(), nil, nil))
 }