+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 * limited session lifetime to 24 hours
 * fixed `data-dev` attribute in `pirsch-events.js`
 * fixed `SessionMaxAge` being used for the maximum session lifetime
+* updated User-Agent blacklist
+* updated referrer blacklist
+* updated referrer mapping
 * updated dependencies
 
 ## 4.1.0